requirements.txt
fastapi==0.129.2
Pillow==12.1.1
pydantic==2.12.5
transformers==4.49.0
fastapi[standard]
pytest